+++ extensions/model-context-protocol-registry/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Model Context Protocol Registry Changelog
 
+## [Add Apify MCP Server] - 2025-05-16
+Add the official Apify MCP server to the Model Context Protocol Registry. Apify MCP server allows AI agents to use 5,000+ ready-made Actors for use cases such as extracting data from websites, social media, search engines, online maps, and more.
+
+## [Add Nuxt MCP Server] - 2025-05-15
+
+Add Official Nuxt MCP Server to registry, this allow docs and modules search.
+
 ## [Chore: Fixed broken MCP links] - 2025-05-09
 
 ## [Initial Version] - 2025-05-09